codebase restructure: more modules added; scanner renamed to tokenizer; token.rs, Expression and Statement moved to /common
