+++ src/common/token_type.rs
@@ -1,22 +1,3 @@
-use std::fmt;
-
-pub struct Token {
-    pub kind: TokenType,
-    pub line: usize,
-}
-
-impl Token {
-    pub fn new(kind: TokenType, line: usize) -> Self {
-        Self { kind, line }
-    }
-}
-
-impl fmt::Debug for Token {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "Token {{ kind: {:?}, line: {:?}}}", self.kind, self.line)
-    }
-}
-
 #[derive(Debug, Clone, PartialEq)]
 pub enum TokenType {
     LeftParen,  // (
