+++ src/runtime/values.rs
@@ -1,0 +1,8 @@
+#[derive(Debug, Clone)]
+pub enum RuntimeValue {
+    Number(f64),
+    String(String),
+    Boolean(bool),
+    Null,
+    Undefined,
+}
