--- src/scanner.rs
@@ -1,8 +1,0 @@
-pub mod scanner;
-pub mod token;
-
-pub use scanner::Scanner;
-pub use token::Token;
-
-#[cfg(test)]
-mod tests;
