+++ src/tokenizer.rs
@@ -1,0 +1,5 @@
+pub mod core;
+#[cfg(test)]
+pub mod tests;
+
+pub use core::Tokenizer;
